requirements.txt
boto3
botocore
python-dotenv
pika
PyMuPDF==1.26.4
pdf2docx==0.5.8
